% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_for_sankey.R
\name{df_for_sankey}
\alias{df_for_sankey}
\title{Title}
\usage{
df_for_sankey(x, y, z)
}
\arguments{
\item{x}{\itemize{
\item data table
}}

\item{y}{\itemize{
\item source column
}}

\item{z}{\itemize{
\item target column
}}
}
\description{
Title
}
