% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KZZLcharts-package.R
\docType{package}
\name{KZZLcharts-package}
\alias{KZZLcharts}
\alias{KZZLcharts-package}
\title{KZZLcharts: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}: 

Authors:
\itemize{
  \item Jakub Miszczyszyn (\href{https://orcid.org/0009-0002-8592-946X}{ORCID}) [contributor]
}

}
\keyword{internal}
